A few bug fixes in the vcgen for patterns.
Added a tag field to the Aggr term to support datatypes and tagged unions.
New term tagof(-) extracts the tag of an aggregate.  Modified pattern
matching and treatment of e.f and e->f for tagged unions to use the
tagof(-).

Vcgen now builds terms for datatypes, structs, and unions.  Unions are
treated as if they were always tagged.  So in the term language, you
can only reduce t.f when f was the last member written in the union
(otherwise, the term remains abstract.)