Got rid of relations for function types -- we use assertions now. This changed absyn.h which is why everything else changed.
Added support for @throws clauses on functions.  For example, one
can say @throws(0) and vcgen will make sure that the function
doesn't throw exceptions.  This subsumes the "no_throw" attribute
(though I haven't yet taught the flow analysis about this.)

Better yet, you can write something like:

`a hd(list_t<`a> x) @throws(x == NULL);

This tells us that the function only throws an exception when the
input list is NULL.  So for instance, you can write:

int f(list_t<int> y) @throws(0) {
  if (y == NULL) return 0;
  return hd(y);
}

There's still no way to name what exceptions might be thrown
since I need to add a way for an assertion to claim that the
tag of a datatype value is equal to some particular constructor.

Reworked unification and subtyping to use the new prover when
checking that the requires/ensures/throws clauses of two function
types are related.  This uncovered a bug in the constraint prover
that we need to track down.

Reworked the way substitution and copying happen for function types to
make sure we respect the binding of parameters (and the return- value)
in @requires, @ensures, and @throws clauses.  What we had before
wasn't really right.  The invariant we maintain is that the
requires_assn, ensures_assn, and throws_assn for a FnInfo always refer
to the arg_vardecls and return_value for that same FnInfo.  Since the
types of these vardecls may change upon substitution, we may have to
generate new vardecls and update these assertions.  Note that the
requires_clause, ensures_clause, and throws_clause (which are
expressions, not assertions) are ignored after kind-checking since at
that point, we convert them to the associated assertions.