Bug fixes in vcgen for datatype patterns -- this caused a lot of the changes in the bin/genfiles.
Bug fix for continue in vcgen.

Bug fix in assndef for substitution in proj terms.

There's a bug in toc.cyc that I haven't yet tracked down.  See
tools/bison/lalr.cyc:620 which has the god awful expression
*temp_R[*sp++]++ = (short)i.  With the changes below, vcgen managed to
get rid of a null or bounds check on this which seemed to trigger a
case that we've never run into before.  I worked around this for now
by rewriting the expression, but we need to fix this.

Added some @throws(false) clauses to a few library functions.  (Should
probably do more.)

Fixed some performance bugs in assndef.  I'm close to being able to
turn --allpaths on by default.  Only a few files with lots of nested
if-then-else's (e.g., cyclone.cyc) blow up.

One performance bug in assndef had to do with the way I was memoizing
reductions -- I kept inserting fresh nodes into the hashtable,
assuming that this would replace old nodes, but of course it doesn't.

I did some profiling and determined that our hash functions for
assertions and terms really sucked.  First, using xor for addresses is
a bad idea -- most of the bits cancel out.  Using addition works a lot
better.  Second, shifting out the least significant bits is important
and tricky since the allocator tends to at least double-word align
things, and for larger objects, tends to align things on larger
boundaries.  For now, I'm shifting out the least 4 significant bits
and this seems to be the right value for reducing collisions.

A few other changes had drastic improvements like comparing the second
operand of an and(-,-) or or(-,-) first, instead of second (since we
tend to build them in a particular way that makes the second one
shallower.)

We now do a little factoring on assertions generated by pattern
matching.  At issue is when we have an Aggr_p or Datatype_p such as
(p_1,p_2,...,p_n) and our input assertion is A, then we end up
generating for the successful match an assertion of the form
A & p_1 & p_2 & ... & p_n, but for the failure case, we were generating:

(A & !p_1) || (A & p_1 & !p_2) || (A & p_1 & p_2 & !p_3) || ... ||
(A & p_1 & ... & p_n-1 & !p_n).

Now we end up generating the more compact tree:

A & (!p_1 || (p_1 && (!p_2 || (p_2 && (!p_3 || (...(p_n_1 && !p_n)...))))))

where all of the common conjuncts are factored out.  This turned out
to be crucial for complicated patterns with --allpaths turned on.  We
might want to do some more simplification of these assertions in the
future...

Modified vcgen to keep track of the exception value that is thrown, at
least in some cases, though there's no way yet for the user to specify
this in a @throws clause.

Added a new pragma print_dag which prints out the dag representation
of an assertion (print_assn prints out the reduced tree.)