Previously we had a hack to change
  ( (t)f )(...)

to

  ({(t)f})(...)

in genfiles to make gcc 3.4 happy.  This doesn't satisfy
gcc 4.0 so we use

  ({ t g = (t) f; g; })(...)

now.

I've bootstrapped under gcc 3.4.2 and gcc 4.0 in Linux.
This revealed a small bug in absyn.cyc (a static declaration
that should not have been static).

I find gcc 4.0 has a new warning we are tripping over,

  pointer targets in initialization differ in signedness

which seems to come from strings.  These are compiled
in genfiles to arrays of chars, and pointers to them are
put into struct _fat_ptr which expects pointers to unsigned
chars.  Annoying but I didn't do anything about it.