Fixed a bug in flow analysis.  If you declared a datatype field that was only carrying bits-only values, but didn't initialize it, then the flow analysis still considered the value to be fully initialized (even though the tag was not yet written.)  So I added an extra dummy field corresponding to the tag and made it uninitialized.  The only way to initialize it is to write the full datatype constructor into the location.
Added a new primitive tagof(e).  Here, e must be a pointer to a
datatype or datatype field.  It returns the tag of the datatype value
as an unsigned integer.  This is useful for doing quick comparisons
between two datatype values (which we need in termcmp for vcgen.)  If
we had some way to "cast" a datatype to a particular field and to
extract the members of a datatype field, then this would give us the
facilities needed to eliminate datatype pattern matching.

Modified termcmp, expcmp, and typecmp functions to use tagof to
rule out the cases where the top-level constructors are different.
This is a huge savings for the vcgen stuff since we do a lot of
term comparisons.  For instance, when compiling getopt.cyc we do
more than 200 million comparisons and this takes up 40% of the
compile time according to gprof.  Using tagof(-) instead of pattern
matching shaved a full 2 seconds (out of 8) off of the compile time.