* Major bug fix for function call.  We weren't updating the exception handler, nor considering the noreturn attribute.
* Bug fix for addition, increment on zeroterm pointers -- need to update
the exception handler.

* Improved treatment of bounds checks for thin pointers.  We were only
allowing the check to be eliminated when we could prove the index was
less than the known bound from the type.  But we may be able to more
generally prove that the index is in bounds from other information
(e.g., a @requires clause.)

* Fixed some of the accounting for null checks and bounds checks
when PERFORMANCE_DEBUG is defined.

* Improved vcgen error messages so that we include the context's varmap.

* Improved parentheses for assertion-to-string function.