Augmented graphs in pratt_prover.cyc with a dictionary mapping terms to graph nodes for faster lookups.  Before, we had a sorted list (which we still need to keep around for iterating through the nodes of the graph) but this made lookups linear.  Now they are at least logarithmic.
This drastically cut the running time of the compiler when compiling
Trevor's indexer.cyc.  Before, it was taking 48s on this particular
machine, and now it takes 14s.  We could probably improve things a bit
more if we augmented the dictionary code to make it possible to find
the least term less than another term (to make it easier to insert a
new node) since that's still a linear-time operation.