Many small changes to vcgen, some of which were bug fixes that, once fixed, slowed things down a lot.  So Trevor, your code takes a lot longer than it used to (29s on my laptop instead of 15s), but for apparently good reasons.  We will continue to hammer on getting this down to a reasonable time.  Most of the time is spent in the constraint solver, so we are going to re-think how we are doing this to improve performance and increase precision.
* Added some functions to sets and hashtables.

* Some more changes to allow memory to be mentioned in assertions.
You can now put @requires/@ensures/@throws clauses on functions that
do dereferencing.  Currently, the @requires clause is a predicate on
the memory right before the function call, whereas the @ensures and
@throws clauses are predicates that hold on the memory right after the
function returns/throws.  In the long term, we'll need to provide
notation for mentioning the old memory in the @requires/@ensures
clauses as well.  [Note:  Yanling is still working on a fix for
the @ensures clause...]

* Some processing to vcgen to not treat assertions (@requires/@ensures
/@throws/@assert/@assert_false) as if they'll be executed.  In particular,
a dereference does not require the bounds to hold, can't throw an
exception, etc.  This is sound only because in the model, memory
is a total function on addresses (including NULL).

* Changed offsetf to take an integer representing which field instead
of using the field name.  Should improve hashing/sharing and so forth.

* More reductions in select(update(m,a1,v),a2) --> select(m,a2) by
leveraging information about when we can trivially show a1 != a2.
In particular, if a1 and a2 are both addresses of variables, or
malloc points, etc. then we can determine that when they aren't
equal.

* Fixed a few performance problems with term comparison and hashing.

* Switched from dictionaries to hash tables for mapping terms to nodes
within a graph (also fixing a bug in the lookup).