Modified primops so that we can distinguish signed and unsigned versions of Div, Mod, Lt, Gt, Lte, and Gte.  We always build abstract syntax with the signed version, and the type-checker rewrites to the unsigned version when it determines that something is unsigned.  We needed this to be able to do constant folding within vcgen for Div and Mod.
Fixed a bug in the subst_close_graph code.  We were supposed to
substitute better terms throughout the graph, but were never
substituting integer constants (since, except for 0, none of them
appear as explicit terms within the graph.)

Fixed a bug in eq_node_const.  The code was checking to see if there
is a constant c such that t-0 <= c and 0-t <= c to determine that
t==c.  But this only holds when t==0.  The proper check is to see
when t-0 <= c and 0-t <= -c.  Then we can conclude t==c.

Tried to improve the synthesis of loop invariants.  What I do is
calculate the set of program variables that might change in the body
of the loop.  If a program variable x changes, then I introduce a
fresh logic variable v for x in the assertion map (and add
v==oldmap(x) to the assertion.)  This ensures that any term which
happens to be loop invariant does not have a logic variable in it that
we might rename as a result of unifying the old and new assertion
maps.  The old approach did not introduce this extra level of
indirection.  As a result, if we had {#mem = v, x=sel(v,...)} &&
sel(v,...) < k and the loop changes memory, but not x, then we would
end up with an invariant of the form {#mem = v, x=w} && true.  With
the new approach, we start off with {#mem = w, x=sel(v,...)} && w=v &&
sel(v,...) < k and then the loop invariant becomes {#mem = w,
x=sel(v,...)}  && sel(v,...) < k.