Rerun ./configure
** Many updates to the the inference mode. cyclone-inf can now handle
all the code in the compiler. I didn't bootstrap the compiler using
inference because this requires protecting library code with
__noinference__ blocks, or coming up with interface files for
libraries after inference. Ignoring the libraries, cyclone-inf is able
to infer fat/thin pointer bounds annotations for all files in src.

You can try this out with "make cyclone_src_inf". The results of
inference get put in build/boot/inf/.cych and
build/boot/inf/.cyco. Once that's done you can run "make
src_inf_stats" to get a summary of the results of inference.

The short story is that there are 1408 pointer bounds expressions that
appear on types that are not purely local to a function body. Of
these, there are 14 pointers that were annotated as thin but are
inferred as fat. This difference is because of the way zeroterm
interacts with fat/thin -- we can get away with using a thin pointer
when the array is zerotermed, but the inference isn't always smart
enough to figure this out and chooses fat instead. There are 11
pointers that were annotated as fat but were inferred to be thin --
most of these are because some library code that wasn't analyzed and
wasn't protected with __noinference__ requires the pointer to be fat.

Some of the changes related to this were to prevent pointer bounds
from being constrained prematurely. For instance, in tcCmpBinop we
would occasionally unify types eagerly, unnecessarily constraining
pointer bounds expressions. This unification is now delayed, some
times until code generation. This explains the changes to several of
the genfiles (printf, scanf etc.) -- something like

        if(p == NULL)

used to get compiled to

        if(p.curr == _tag_fat(0,0,0).curr)

but now gets compiled to

        if(p.curr == 0)


** I was stupidly allowing subtyping under references with respect to
effect annotations -- e.g int *`r1 * <: int *`r1+`r2 *, which is
clearly no good when `r2 does not outlive `r1. Got rid of this bug,
and cleaned up subtype.cyc while doing so.

There was some code in tcdecl, lex and scheme-2 which used to use
outlives that actually relied on this bug to typecheck. The way in
which injections were being typechecked relied on this bug too. So all
of that has changed.