Several fixes for bugs uncovered when getting the benchmarks up to date.
The changes to the genfiles are mostly because of a fix for the following bug:

The typechecker and evexp were treating bit-shift operations as always
being unsigned. For right shifts this is not consistent with C
semantics. I've changed this around so that bit shifting takes the
sign into account and doesn't just coerce everything to unsigned. The
changes to genfiles are mostly because of the removal of these
coercions.

Other bugs include a problem with the way injections were chosing
which datatypefield to use -- looks like this bug's been in the
compiler for a long time. printf(" %d ", 0), would result in a runtime
error of "Invalid argument" because the 0 was being promoted to NULL
and then to char ?.

Also a change to libc.cys for sys/wait.h, which used __typeof instead
of typeof in some macros.