Rerun ./configure.
These changes are intended to make it possible for me to compile
the output of the compiler with Microsoft's Visual C/C++ compiler
or hopefully, any other C compiler.  I'm not yet done since VC++
demands more casts than GCC, but I'm pretty close.  This code
has been tested on the compiler and tests/ but nothing else.

Added a new file, bin/cyc-lib/vc_cyc_include.h which should be
used instead of cyc_include.h when compiling for VC++.  Basically,
this just removes all of the macros which used statement-
expressions in favor of static, inlined functions.

Made it an error to initialize a [tagged] union with anything but
the first member.  This reflects a constraint put on us by vanilla
C which provides no way to initialize unions at top-level except
by using the first member.  (Ugh.)  This forced me to change a
few things here and there in the compiler where we violated the
constraint.  (I did have a way to deal with this, but it involved
an ugly C++ hack which was too hard to get right, so I opted for
the constraint instead.)

New phase in backend, lower.cyc, takes care of eliminating statement-
expressions and sequentializing the code.  This phase essentially
A-normalizes the output of toc.cyc, creating temps for intermediate
expressions, eliminates things like +=, ++, &&, ||, etc. in favor of
simpler constructs, uses only "if (...) goto L else goto L" and "goto
L" for control-flow, etc.  It's low enough that you could almost feed
it into a generic compiler IR.

To support this, I had to modify toc and remove_aggregates extensively
to add type information on all of the generated syntax.  I think this
was a good move anyway because it forced me to clear up lots of
niggling typing issues in the back-end, and helped to eliminate some
unnecessary casting (as well as insert some necessary casting.)
I'm also hoping these changes will make it easier to fix the problem
with GCC 4.x.

Modified pattern match compiler so that it doesn't always compute
the complete path from the root to a variable being extracted for
each variable in a pattern.  Avoids generating too many temps.

The rest of the changes were necessary to get this to scale in
a reasonable fashion.  For instance, I had to modify absyndump.cyc
so that it was effectively tail-recursive on sequences of
statements or variable declarations since lowered parse_tab.cyc
is simply tremendous.  I also modified the match compiler so that
it caches intermediate pointers in temps so that we aren't always
recomputing the path from the root being matched to the leaves
corresponding to the variable patterns.

These changes didn't really slow down our portion of the compiler,
but they did slow down GCC's portion by a fair amount, mostly
because our output is a lot bigger.