Bug fix in insert_checks.  It was running into valueof(T) but not expecting to.  The reason it might encounter it is due to the presence of @assert's.
A bit of cleanup in vcgen for rhs-exps.  We were doing a hack
to evaluate constant expressions early, but too many compound
expressions were being treated as symbolic "constants", resulting
in the prover having to treat them abstractly.  A side effect
of fixing this is that fewer expressions are reduced to a
constant, which is why so many of the files changed.

Enhancement for vcgen -- if you declared an array, say char x[255],
then vcgen only discovered that numelts(x) >= 255 when it could
discover numelts(x) == 255.

Bug fix in termcmp for binops.  Wasn't detecting different binops!

Performance fix for Dijstra shortest-paths algorithm.  We were
iterating through all of the nodes in the graph when it was
cheaper to just keep an explicit queue of the ones that needed
to be processed for a given node.  To really make this scale well,
we really ought to replace the naive queue that I'm using with
a Fibonacci heap, but I'm too lazy to do this now.

Better hash functions in assndef.cyc.  I was just doing some
random bit twiddling to stuff together two 32-bit numbers into
a hash code and was afraid that we were getting too many
collisions.  So I adapted http://burtleburtle.net/bob/c/lookup2.c
which is supposed to be a good hash function.  However, I
didn't see much improvement over what I had before.  I also
fiddled with the resizing code in the Hashtable library to
avoid growing the hashtable too aggressively.

Optional enchancement for vcgen --better-widen:

Widening is used to construct loop invariants, by reducing an
assertion A1 || A2 to an assertion A implied by A1 and A2, such that A
is no bigger than A1 or A2.  The default approach is to just intersect
the set of common (positive) literals for A1 and A2.  The flag
--better-widen goes a step further and also adds any literals from A1
that can be proven by A2 using the constraint solver.  (The asymmetry
is there since A1 is typically the proposed loop invariant, and we
want to preserve as much information from that as possible.) This
picks up a number of facts that are otherwise dropped on the floor,
since the constraint solver does substitution, reduction, and
transitive closure of relations that the default widening procedure
cannot.  For example, our default procedure cannot prove the
following code needs no check, whereas --better-widen can:

  int buf[10];

  int n = 10;
  for (int i = 0; i < n; i++)
    buf[i] = i;

In the default mode, the loop invariant generated for this is
simply "n == 10", when we want it to be "n== 10 && i >= 0".
In particular, we need i >= 0 plus the guard to eliminate the
bounds check.  To see what goes wrong, consider the first
widening step after we've generated the initial assertion:
   A0 == (i S<= 0 && 0 S<= i && n == 10)
and the assertion right after the first iteration:
   A1 == (i S<= 1 && 1 S<= i && n == 10)
When we widen this, the only common literal is "n == 10" so
that's the best the loop invariant ever could be.  With the
flag turned on, however, we also are able to show that
0 S<= i is true.  Thus, our candidate loop invariant becomes
"0 S<= i && n == 10".  Note that if you simply declared n
to be an "unsigned int" then the problem goes away since then
we have as part of the loop guard "i U< n" which, coupled
with "n == 10" is enough to establish the safety of the subscript.

Unfortunately, the improved widening picks up too much information,
resulting in excessively long compile times for some files (but
not others.)  For instance, compiling toc.cyc without the flag
takes 15s, but turning the flag on takes 1m18s.

The problem is that we get bigger assertions, and that translates into
more time in the constraint solver (an order of magnitude more time
for some files.)  So I've turned this off by default.  However,
I would recommend turning it on for any benchmarks, especially
ones ported from C.  I'm going to keep re-thinking the constraint
solver to see if we can't improve it so that we can turn this on
by default.