* Fixed a bug where the lowering phase was always run, even if flags such as -stopafter-tc where given.
* Added support for a new type constructor:  @subset(T x | e[x]).
This is a subset-type or comprehension type corresponding to the
set of all T values x, such that e[x] is true.

For example, @subset(int x | x >= 0) is the type of all non-negative
integers, @subset($(int,int) p | p.f0 > p.f1) is the set of all
pairs of integers such that the first component is greater than
the second, etc.  We could define not-null types as:

  typedef @subset(`a *`r x | x != 0) ptr_t<`a,`r>;

The type-checker automatically promotes subset types to their underlying
type.  So, for instance, you can write:

  typedef @subset(int x | x >= 0) nonneg_t;

  int inc(nonneg_t x) { return x+1; }

Here, x is promoted from the subset type to int.  Promotion is
surpressed when the context makes it apparent that a subset type is
wanted.  For instance:

  struct Pair { nonneg_t x,y; };

  struct Pair pair(nonneg_t x) {
    return Pair{x,x};
  }

Here, because the types of the fields are subset types, the uses
of x are not promoted.

Casts from an underlying type to a subset type are also supported.
However, vcgen must be able to prove that the value has the desired
property.  For instance:

  struct Pair pair2(int x) {
    if (x >= 0)
      return Pair{x,x};
    else
      return Pair{0,0};
  }

will type-check, without the test, it won't.  (In the near future,
I plan to insert a run-time test if we can't prove the assertion,
at least when the assertion can be realized as a run-time check.)

Pointers into the middle of a comprehension type are restricted
to read-only.  So for instance, if we have:

  typedef @subset(struct Pair p | p.x >= p.y) ordered_pair_t;

  nonneg_t g(ordered_pair_t p) {
    let xp = &p.x;
    return *xp;
  }

then this succeeds, but if we try to assign to the contents of xp,
we'll get a type-error.  The reason is that of course, if we change
the x component, we might break the property that it's no longer
greater than or equal to the y component.  This is similar to Dan's
pointer under an existential --- however, we can always coerce a
@subtype(T x | e) to a read-only T without having to worry about some
form of scope-extrusion.  This could be potentially generalized: if
the assertion does not mention the component, we could make the
pointer read-write.  If the assertion *only* mentions the component,
then we can push the subset constraint into the pointer element type.
For now, I'm keeping it simple.

Finally, assertions in @subset type constructors are limited to
talking about values that are independent of the contents of
memory.  For instance, the following type is ill-formed:

  @subset(int @x | *x >= 0)

We can only enforce this constraint if we ensure that all updates
via aliases respect the invariant.  Of course, if the pointer
is non-aliasable (or strong-read-only), then we could support this,
but I haven't gotten around to this yet.

This undoubtably has introduced bugs where we were expecting a
particular type-constructor, but now could have a subset type
(e.g., in the back-end, when translating an e.f expression, e
could have a subset type instead of an aggregate type.)  I've
tried to track down as many of these as possible, but let me know
if you run into other problems.